+++ changes.md
@@ -5,6 +5,7 @@
 - Fixed inconsistent response when returning an empty paginator and using transformers.
 - Fixed bug with protected endpoints becoming unprotected when setting protection on both group and route.
 - Fixed bug where routes defined before the replacement of the bound router would be lost.
+- `AuthFilter` now require authentication for internal requests.
 
 #### Added